+++ app/src/main/java/ch/abwesend/privatecontacts/infrastructure/repository/androidcontacts/repository/AndroidContactSaveRepository.kt
@@ -1,4 +1,4 @@
-package ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts
+package ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.repository
 
 import ch.abwesend.privatecontacts.domain.model.contact.IContactIdExternal
 import ch.abwesend.privatecontacts.domain.model.result.ContactDeleteResult
