+++ app/src/test/java/ch/abwesend/privatecontacts/infrastructure/repository/androidcontacts/AndroidContactRepositoryTest.kt
@@ -9,6 +9,7 @@ package ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts
 import android.content.Context
 import ch.abwesend.privatecontacts.domain.model.contact.ContactIdAndroid
 import ch.abwesend.privatecontacts.domain.service.interfaces.PermissionService
+import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.repository.AndroidContactLoadRepository
 import ch.abwesend.privatecontacts.testutil.TestBase
 import ch.abwesend.privatecontacts.testutil.databuilders.someAndroidContact
 import com.alexstyl.contactstore.Contact
